+++ .gitignore
@@ -1,3 +1,1 @@
 node_modules
-/dist
-.DS_Store*
