--- public/constants.js
@@ -1,7 +1,0 @@
-const LOCAL_HOST = "http://localhost:8080";
-const MULTIPLIER = 1; // Sets how much the data should be scaled by.
-
-export {
-    LOCAL_HOST,
-    MULTIPLIER
-};
