+++ public/scripts/generators/data_generator.js
@@ -225,9 +225,3 @@ function create_condition_dataset(dataset, condition_data){
 
   return condition_dataset;
 }
-
-export {
-  get_data,
-  get_data_subset,
-  create_condition_dataset
-}
