+++ public/scripts/generators/gaussian_distribution_generator.js
@@ -1,8 +1,9 @@
+
 const TRIAL_LIMIT = 250;
 
 function generateDistribution(correlation, error, size, numsd, mean, sd){
 
-  // dynamicallyLoadScript(MATHJS_URL);
+  dynamicallyLoadScript(MATHJS_URL);
   var coordinates = {x_values: [], y_values: []};
   var overshootSize = size + 20; // Generating > size will guarantee we have
                                  // a distribution of the speciied size later. 
@@ -26,7 +27,7 @@ function generateDistribution(correlation, error, size, numsd, mean, sd){
   // so that origin as at the top left (this makes the distribution negative instead
   // of positive). To force it to be positive, we flip the whole distribution 
   // across y axis. 
-  for (let i = 0; i<coordinates.y_values.length; i++) {
+  for (i = 0; i<coordinates.y_values.length; i++) {
     coordinates.y_values[i] = coordinates.y_values[i] * (-1);
   }
 
@@ -38,7 +39,7 @@ function initializePoints(coordinates, correlation, size, numsd, mean, sd){
   var x2Val;
   var yVal;
 
-  for (let i = 0; i < size; i++) {
+  for (i = 0; i < size; i++) {
     do {
         xVal = random_bm();
         x2Val = random_bm();
@@ -81,7 +82,7 @@ function readjustPoints(coordinates, correlation, error, size, numsd, mean, sd) 
   var temp_coordinates = {x_values: [], y_values: []};
   var max_iterations = 500;
 
-  for (let i = 0; i<size; i++) {
+  for (i = 0; i<size; i++) {
     if (pointNotWithinRequiredStdDevs(coordinates.x_values[i], coordinates.y_values[i], numsd, mean, sd)){
       coordinates.x_values.splice(i, 1);
       coordinates.y_values.splice(i, 1);
@@ -100,11 +101,8 @@ function readjustPoints(coordinates, correlation, error, size, numsd, mean, sd) 
         if (i > 0) {
           temp_coordinates.x_values = coordinates.x_values[i-1];
           temp_coordinates.y_values = coordinates.y_values[i-1];
-
-          // redefining undefined?
-          // CAITLINNNNNN
-          // temp_coordinates.x_values[i] = x;
-          // temp_coordinates.y_values[i] = y;
+          temp_coordinates.x_values[i] = x;
+          temp_coordinates.y_values[i] = y;
         }
 
         if (!pointNotWithinRequiredStdDevs(x, y, numsd, mean, sd) && (i > 0) &&
@@ -130,10 +128,9 @@ function readjustPoints(coordinates, correlation, error, size, numsd, mean, sd) 
   //}
 }
 
-// CAITLINNNNNNNN
-// function pointNotWithinRequiredStdDevs(x, y, numsd){
-//   return pointNotWithinRequiredStdDevs2(x, y, numsd, 0, 1);
-// }
+function pointNotWithinRequiredStdDevs(x, y, numsd){
+  return pointNotWithinRequiredStdDevs2(x, y, numsd, 0, 1);
+}
 
 function pointNotWithinRequiredStdDevs(x, y, numsd, mean, sd){
 
@@ -164,7 +161,7 @@ function transformPoints(coordinates, mean, sd) {
 
 // Scales the points in a distribution with the given x and y scale values.
 function scalePoints(coordinates, scaleX, scaleY) {
-  for (let i = 0; i<coordinates.x_values.length; i++) {
+  for (i = 0; i<coordinates.x_values.length; i++) {
     coordinates.x_values[i] = coordinates.x_values[i] * scaleX;
     coordinates.y_values[i] = coordinates.y_values[i] * scaleY;
   }
@@ -172,7 +169,7 @@ function scalePoints(coordinates, scaleX, scaleY) {
 
 // Translates the points in a distribution with the given x and y translate values.
 function translatePoints(coordinates, x, y) {
-  for (let i = 0; i<coordinates.x_values.length; i++) {
+  for (i = 0; i<coordinates.x_values.length; i++) {
     coordinates.x_values[i] = coordinates.x_values[i] + x;
     coordinates.y_values[i] = 1 - (coordinates.y_values[i] + y);
   }
@@ -209,7 +206,7 @@ function getPearsonCorrelation(x, y) {
   var sum_x2 = 0;
   var sum_y2 = 0;
 
-  for(let i=0; i< shortestArrayLength; i++) {
+  for(var i=0; i< shortestArrayLength; i++) {
       sum_x += x[i];
       sum_y += y[i];
       sum_xy += xy[i];
@@ -239,9 +236,4 @@ function dynamicallyLoadScript(url) {
   script.src = url; // Set it's src to the provided URL
 
   document.head.appendChild(script); // Add it to the end of the head section of the page (could change 'head' to 'body' to add it to the end of the body section instead)
-}
-
-export {
-  TRIAL_LIMIT,
-  generateDistribution
 }
