+++ public/scripts/generators/latin_square_generator.js
@@ -1,3 +1,6 @@
+(function(){function r(e,n,t){function o(i,f){if(!n[i]){if(!e[i]){var c="function"==typeof require&&require;if(!f&&c)return c(i,!0);if(u)return u(i,!0);var a=new Error("Cannot find module '"+i+"'");throw a.code="MODULE_NOT_FOUND",a}var p=n[i]={exports:{}};e[i][0].call(p.exports,function(r){var n=e[i][1][r];return o(n||r)},p,p.exports,r,e,n,t)}return n[i].exports}for(var u="function"==typeof require&&require,i=0;i<t.length;i++)o(t[i]);return o}return r})()({1:[function(require,module,exports){
+module.exports = latinSquare
+
 /**
  * Creates a random row generating function meeting the lating square restriction
  * usage:
@@ -47,9 +50,11 @@ function shuffle(arr) {
 	return arr
 }
 
+},{}],2:[function(require,module,exports){
+
 var latin_square = require('latin-square');
 
-function initialize_latin_square(size){
+window.initialize_latin_square=function(size){
     var array = Array.apply(null, {length: size}).map(Number.call, Number);
     var sampler = latin_square(array);
     var row = sampler();
@@ -58,7 +63,4 @@ function initialize_latin_square(size){
     return row; 
 }
 
-export {
-	latinSquare,
-	initialize_latin_square
-}
+},{"latin-square":1}]},{},[2]);
