+++ public/views/index.html
@@ -16,6 +16,44 @@
   </head>
 
   <body>
-    <div id="sparoot"></div>
+    <div class = "box">
+      <img src="../img/VCL_lab_logo.png"></img>
+
+      <h1>Proof of Concept Experiments</h1>
+      <br>
+        <div class="btn-group">
+
+          <button class="btn btn-info btn-lg dropdown-toggle" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" style="width: 250px">
+            JND
+          </button>
+
+          <div class="dropdown-menu" aria-labelledby="dropdownMenu2" style="width: 250px">
+            <h6 class="dropdown-header">Foundational</h6>
+              <a class="dropdown-item" href="/jnd_scatter">Scatter : Base</a>
+              <a class="dropdown-item" href="/jnd_strip">Strip : Base</a>
+            <h6 class="dropdown-header">Design</h6>
+            <a class="dropdown-item" href="/jnd_scatter_distractor_rainbow">Scatter: Distractor Rainbow</a>
+              
+          </div>
+
+        </div>
+
+        <div class="btn-group">
+
+          <button class="btn btn-info btn-lg dropdown-toggle" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" style="width: 250px; margin-left: 20px">
+            Stevens
+          </button>
+
+          <div class="dropdown-menu" aria-labelledby="dropdownMenu2" style="width: 250px">
+            <h6 class="dropdown-header">Foundational</h6>
+              <a class="dropdown-item" href="/stevens_scatter">Scatter: Base</a>
+              <a class="dropdown-item" href="/stevens_strip">Strip : Base</a>
+           <!--  <h6 class="dropdown-header">Design</h6> -->
+          </div>
+
+        </div>
+
+    </div>
   </body>
+
 </html>
